Complete page en complete component
Scrappy ik heb de complete page gescheiden zodat we geen searchparams errors meer krijgen en een nieuwe component aangemaakt voor checkout-complete-content.tsx
